fix(discord-voice): logconf.h dependency regression
